`refactor`: moved tests from root to `tests/`; `fix`: indexing bug in find
