+++ src/charset.rs
@@ -133,110 +133,3 @@ impl CharSet {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_new_empty_full_charset() {
-        let empty = CharSet::new();
-        let full = CharSet::full();
-        for i in 0..=255 {
-            assert!(!empty.contains(i));
-            assert!(full.contains(i));
-        }
-    }
-
-    #[test]
-    fn test_add_byte_charset() {
-        let mut set = CharSet::new();
-        assert!(!set.contains(b'a'));
-        set.add_byte(b'a');
-        assert!(set.contains(b'a'));
-        assert!(!set.contains(b'b'));
-    }
-
-    #[test]
-    fn test_add_range_charset() -> Result<()> {
-        let mut set = CharSet::new();
-        set.add_range(b'a', b'c')?;
-        assert!(set.contains(b'a'));
-        assert!(set.contains(b'b'));
-        assert!(set.contains(b'c'));
-        assert!(!set.contains(b'd'));
-        Ok(())
-    }
-
-    #[test]
-    fn test_add_range_invalid_charset() {
-        let mut set = CharSet::new();
-        assert!(matches!(set.add_range(b'z', b'a'), Err(Error::Parser(_))));
-    }
-
-    #[test]
-    fn test_add_class_digit_charset() -> Result<()> {
-        let mut set = CharSet::new();
-        set.add_class(b'd')?;
-        assert!(set.contains(b'0'));
-        assert!(set.contains(b'5'));
-        assert!(set.contains(b'9'));
-        assert!(!set.contains(b'a'));
-        assert!(!set.contains(b' '));
-        Ok(())
-    }
-
-    #[test]
-    fn test_add_class_space_charset() -> Result<()> {
-        let mut set = CharSet::new();
-        set.add_class(b's')?;
-        assert!(set.contains(b' '));
-        assert!(set.contains(b'\t'));
-        assert!(set.contains(b'\n'));
-        assert!(set.contains(b'\r'));
-        assert!(!set.contains(b'a'));
-        Ok(())
-    }
-
-    #[test]
-    fn test_add_class_invalid_charset() {
-        let mut set = CharSet::new();
-        assert!(matches!(set.add_class(b'Z'), Err(Error::Parser(_))));
-        assert!(matches!(set.add_class(b'%'), Err(Error::Parser(_))));
-    }
-
-    #[test]
-    fn test_contains_charset() {
-        let mut set = CharSet::new();
-        set.add_byte(b'x');
-        assert!(set.contains(b'x'));
-        assert!(!set.contains(b'y'));
-    }
-
-    #[test]
-    fn test_invert_charset() {
-        let mut set = CharSet::new();
-        set.add_byte(b'a');
-        set.add_byte(b'z');
-
-        assert!(set.contains(b'a'));
-        assert!(!set.contains(b'b'));
-        assert!(set.contains(b'z'));
-
-        set.invert();
-
-        assert!(!set.contains(b'a'));
-        assert!(set.contains(b'b'));
-        assert!(!set.contains(b'z'));
-
-        let mut full_set = CharSet::full();
-        full_set.invert();
-        let empty_set = CharSet::new();
-        assert_eq!(full_set.bytes, empty_set.bytes);
-
-        let mut empty_set_2 = CharSet::new();
-        empty_set_2.invert();
-        let full_set_2 = CharSet::full();
-        assert_eq!(empty_set_2.bytes, full_set_2.bytes);
-    }
-}
